+++ jest.config.js
@@ -1,3 +1,4 @@
 module.exports = {
-  testEnvironment: "node"
+  testEnvironment: "node",
+  testRunner: "jest-jasmine2"
 }