+++ spec/form.test.js
@@ -1,13 +1,12 @@
 'use strict';
-const  {URL} = require('url'),
-  {newAgent} = require('../lib/mechanize/agent'),
+const {newAgent} = require('../lib/mechanize/agent'),
   {newPage} = require('../lib/mechanize/page'),
   {newButton} = require('../lib/mechanize/form/button'),
   {fixture} = require('./helpers/fixture.js'),
   {mockServer} = require('./helpers/mock_server.js');
 
 describe('Mechanize/Form', () => {
-  let server, baseUrl, host, domain, agent, form;
+  let server, baseUrl, host, agent, form;
 
   beforeAll(done => {
     server = mockServer();
@@ -17,16 +16,15 @@ describe('Mechanize/Form', () => {
   beforeEach(() => {
     baseUrl = process.env.SERVER_BASE_URL;
     host = process.env.SERVER_HOST;
-    const url = new URL(baseUrl);
-    domain = url.hostname;
     agent = newAgent();
   });
   describe('with no action attribute', () => {
     beforeEach(() => {
-      let uri, body, page;
-      uri = 'form.html';
-      body = fixture('login_no_action.html');
-      page = newPage({uri, body});
+      const uri = 'form.html';
+      const body = fixture('login_no_action.html');
+      const page = newPage({
+        uri, body
+      });
 
       form = page.form('login');
     });
@@ -45,10 +43,11 @@ describe('Mechanize/Form', () => {
 
   describe('with action attribute', () => {
     beforeEach(() => {
-      let uri, body, page;
-      uri = baseUrl;
-      body = fixture('login.html');
-      page = newPage({uri, body, agent});
+      const uri = baseUrl;
+      const body = fixture('login.html');
+      const page = newPage({
+        uri, body, agent
+      });
 
       form = page.form('MAINFORM');
     });
@@ -87,8 +86,10 @@ describe('Mechanize/Form', () => {
     });
 
     it('should have buildQuery', () => {
-      expect(form.buildQuery()).toEqual([ [ 'userID', '' ], [ 'name', '' ],
-        [ 'street', 'Main' ] ]);
+      expect(form.buildQuery()).toEqual([
+        ['userID', ''], ['name', ''],
+        ['street', 'Main']
+      ]);
     });
 
     it('should have requestData', () => {
@@ -97,7 +98,9 @@ describe('Mechanize/Form', () => {
 
     describe('and adding button to query', () => {
       beforeEach(() => {
-        const button = newButton({name: 'button'});
+        const button = newButton({
+          name: 'button'
+        });
         form.addButtonToQuery(button);
       });
       it('should add button to requestData', () => {
@@ -123,12 +126,9 @@ describe('Mechanize/Form', () => {
     });
 
     describe('then submitting form', () => {
-      let submitPage, submitError;
       beforeEach(done => {
         server.postForm.calls.reset();
         form.submit({})
-          .then(page => submitPage = page)
-          .catch(error => submitError = error)
           .then(() => done());
       });
       it('should post the form', () => {
@@ -136,7 +136,7 @@ describe('Mechanize/Form', () => {
           path: '/Login.aspx',
           headers: {
             'user-agent': jasmine.stringMatching(
-                /Mechanize\/[.0-9]+ Node.js\/v[.0-9]+ \(http:\/\/github.com\/srveit\/mechanize-js\/\)/),
+              /Mechanize\/[.0-9]+ Node.js\/v[.0-9]+ \(http:\/\/github.com\/srveit\/mechanize-js\/\)/),
             accept: '*/*',
             'accept-encoding': 'gzip,deflate',
             'content-type': 'application/x-www-form-urlencoded',
@@ -161,8 +161,10 @@ describe('Mechanize/Form', () => {
       });
 
       it('should not include field in buildQuery', () => {
-        expect(form.buildQuery()).toEqual([ [ 'userID', '' ],
-          [ 'street', 'Main' ] ]);
+        expect(form.buildQuery()).toEqual([
+          ['userID', ''],
+          ['street', 'Main']
+        ]);
       });
 
     });
@@ -186,7 +188,9 @@ describe('Mechanize/Form', () => {
     beforeEach(() => {
       const uri = null,
         body = fixture('form_text_plain.html'),
-        page = newPage({uri, body});
+        page = newPage({
+          uri, body
+        });
 
       form = page.form('form1');
       form.setFieldValue('text', 'hello');