+++ spec/helpers/fixture.js
@@ -3,6 +3,8 @@ const fs = require('fs'),
   path = require('path');
 
 const fixture = filename =>
-  fs.readFileSync(path.join(__dirname, '../fixtures', filename), 'utf8');
+ // eslint-disable-next-line no-sync
+ fs.readFileSync(path.join(__dirname, '../fixtures', filename), 'utf8');
+
 
 exports.fixture = fixture;