+++ spec/helpers/mock_server.js
@@ -9,18 +9,21 @@ const createMockServer = ({name, rootPath, environment, port}) => {
   const serverName = name,
     savedEnvironment = {},
 
+    mock = {},
+
     space = length =>
-      '                                                                 '
-        .substr(0, length),
+    '                                                                 '
+    .substr(0, length),
 
     camelToDash = str => str
-      .replace(/(^[A-Z])/, ([first]) => first.toLowerCase())
-      .replace(/([A-Z])/g, ([letter]) => `-${letter.toLowerCase()}`)
-      .replace(/([a-z])([0-9])/g, ([letter, number]) => `${letter}-${number}`),
+    .replace(/(^[A-Z])/, ([first]) => first.toLowerCase())
+    .replace(/([A-Z])/g, ([letter]) => `-${letter.toLowerCase()}`)
+    .replace(/([a-z])([0-9])/g, ([letter, number]) => `${letter}-${number}`),
 
     toXml = ({name, value, includeDeclaration = false, indent = 0}) => {
       const xml = includeDeclaration ?
-        '<?xml version="1.0" encoding="UTF-8"?>\n' : '';
+        '<?xml version="1.0" encoding="UTF-8"?>\n' :
+        '';
       name = camelToDash(name);
       if (value === null || value === undefined) {
         return xml + space(indent) + '<' + name + ' nil="true"/>';
@@ -40,32 +43,34 @@ const createMockServer = ({name, rootPath, environment, port}) => {
         if (value.length === 0) {
           return xml + space(indent) + '<' + name + ' type="array"/>';
         }
-        return xml + space(indent) + '<' + name + ' type="array">\n' + 
+        return xml + space(indent) + '<' + name + ' type="array">\n' +
           value.map(
             item => toXml({
               name: name + '-item',
               value: item,
-              indent: indent + 2})
-          ).join('\n') + '\n' +
-          space(indent) + '</' + name + '>';
-      } else {
-        if (_.isEmpty(value)) {
-          return xml + space(indent) + '<' + name + '/>';
-        }
-        return xml + space(indent) + '<' + name + '>\n' + 
-          _.map(
-            value,
-            (item, key) => toXml({
-              name: key,
-              value: item,
-              indent: indent + 2})
+              indent: indent + 2
+            })
           ).join('\n') + '\n' +
           space(indent) + '</' + name + '>';
       }
+      if (_.isEmpty(value)) {
+        return xml + space(indent) + '<' + name + '/>';
+      }
+      return xml + space(indent) + '<' + name + '>\n' +
+        _.map(
+          value,
+          (item, key) => toXml({
+            name: key,
+            value: item,
+            indent: indent + 2
+          })
+        ).join('\n') + '\n' +
+        space(indent) + '</' + name + '>';
+
     },
 
     mockHandler = (name, method, path, responseName) => {
-      mock[name] = jasmine.createSpy(name).and.returnValue({}); // eslint-disable-line jasmine/no-unsafe-spy
+      mock[name] = jasmine.createSpy(name).and.returnValue({});
       mock.app[method](rootPath + path + '*', (req, res) => {
         // console.log('called', serverName, 'mock', name, path);
 
@@ -96,13 +101,13 @@ const createMockServer = ({name, rootPath, environment, port}) => {
     },
 
     addDefaultHandler = () =>
-      mock.app.all(
-        '*',
-        (req, res, next) => {
-          console.log('no', serverName, 'handler for', req.method, req.url); // eslint-disable-line no-console
-          next();
-        }
-      ),
+    mock.app.all(
+      '*',
+      (req, res, next) => {
+        console.log('no', serverName, 'handler for', req.method, req.url); // eslint-disable-line no-console
+        next();
+      }
+    ),
 
     setEnvironment = mockUrl => {
       const urlParsed = url.parse(mockUrl);
@@ -124,45 +129,54 @@ const createMockServer = ({name, rootPath, environment, port}) => {
 
     restoreEnvironment = () => _.forEach(
       environment,
-      (value, key) => process.env[key] = savedEnvironment[key]
+      (value, key) => {
+        process.env[key] = savedEnvironment[key];
+      }
     ),
-        start = done => {
-          server = mock.app.listen(port || 0, () => {
-            const mockPort = server.address().port,
-                  mockUrl = 'http://localhost:' + mockPort + rootPath;
-            // console.log(serverName, 'listening at', mockUrl);
-            setEnvironment(mockUrl);
-            done();
-          });
-        },
+    start = done => {
+      server = mock.app.listen(port || 0, () => {
+        const mockPort = server.address().port,
+          mockUrl = 'http://localhost:' + mockPort + rootPath;
+        // console.log(serverName, 'listening at', mockUrl);
+        setEnvironment(mockUrl);
+        done();
+      });
+    },
 
     stop = done => {
       restoreEnvironment();
       server.close(() => done());
-    },
-        
-    mock = {
+    };
+
+  Object.assign(
+    mock,
+    {
       app: express(),
       mockHandler,
       addDefaultHandler,
       start,
       stop
-    };
+    }
+  );
 
-  mock.app.use(bodyParser.urlencoded({ extended: false }));
+  mock.app.use(bodyParser.urlencoded({
+    extended: false
+  }));
   mock.app.use(bodyParser.json());
 
   return mock;
 };
 
-const mockServer = done => {
+const mockServer = () => {
   const environment = {
     SERVER_BASE_URL: '<mockUrl>',
     SERVER_HOSTNAME: '<mockHostname>',
     SERVER_HOST: '<mockHost>'
   };
 
-  const server = createMockServer({name: 'server', rootPath: '', environment});
+  const server = createMockServer({
+    name: 'server', rootPath: '', environment
+  });
   server.mockHandler('postForm', 'post', '/');
   server.mockHandler('getPage', 'get', '/');